+++ Projects/Feature/Ledger/Sources/Scene/Ledger/Sheets/DatePicker/DatePickerReactor.swift
@@ -25,8 +25,7 @@ final class DatePickerReactor: Reactor {
     @Pulse var endDate: DateInfo
     @Pulse var isWarning = false
     @Pulse var destination: Destination?
-    @Pulse var yearList: [Int]
-    @Pulse var monthList: [Int]
+    @Pulse var dateList: DateList
     
     enum Destination {
       case ledger
@@ -37,6 +36,11 @@ final class DatePickerReactor: Reactor {
   enum PickerState {
     case start
     case end
+  }
+  
+  struct DateList {
+    let year: [Int]
+    let month: [Int]
   }
 
   let initialState: State
@@ -52,8 +56,10 @@ final class DatePickerReactor: Reactor {
     self.initialState = State(
       startDate: startDate,
       endDate: endDate,
-      yearList: Array((currentYear - 15)...currentYear).reversed(),
-      monthList: Array(1...12).reversed()
+      dateList: DateList(
+        year: Array((currentYear - 15)...currentYear).reversed(),
+        month: Array(1...12).reversed()
+      )
     )
     self.service = ledgerService
   }
@@ -69,8 +75,8 @@ final class DatePickerReactor: Reactor {
       ])
     case .selectDate(let row, let component):
       switch component {
-      case 0: return .just(.setYear(currentState.yearList[row]))
-      case 1: return .just(.setMonth(currentState.monthList[row]))
+      case 0: return .just(.setYear(currentState.dateList.year[row]))
+      case 1: return .just(.setMonth(currentState.dateList.month[row]))
       default: return .empty()
       }
     case .didTapCompleteButton:
@@ -101,12 +107,12 @@ final class DatePickerReactor: Reactor {
     case .setPickerRow:
       switch state.selectedDateType {
       case .start:
-        let yearRow = state.yearList.firstIndex(of: state.startDate.year)
-        let monthRow = state.monthList.firstIndex(of: state.startDate.month)
+        let yearRow = state.dateList.year.firstIndex(of: state.startDate.year)
+        let monthRow = state.dateList.month.firstIndex(of: state.startDate.month)
         newState.pickerRow = (yearRow!, monthRow!)
       case .end:
-        let yearRow = state.yearList.firstIndex(of: state.endDate.year)
-        let monthRow = state.monthList.firstIndex(of: state.endDate.month)
+        let yearRow = state.dateList.year.firstIndex(of: state.endDate.year)
+        let monthRow = state.dateList.month.firstIndex(of: state.endDate.month)
         newState.pickerRow = (yearRow!, monthRow!)
       }
     case .setYear(let year):
