+++ Projects/Feature/Ledger/Sources/Scene/Ledger/Sheets/DatePicker/Views/MMDatePicker.swift
@@ -1,0 +1,60 @@
+import UIKit
+
+import DesignSystem
+import Utility
+
+import FlexLayout
+import PinLayout
+import ReactorKit
+
+final class MMDatePicker: UIView, View {
+  var disposeBag = DisposeBag()
+  
+  private let yearPicker = MMPicker(type: .year)
+  private let monthPicker = MMPicker(type: .month)  
+  private let rootContainer = UIView()
+  
+  override init(frame: CGRect) {
+    super.init(frame: frame)
+    setupConstraints()
+  }
+
+  @available(*, unavailable)
+  required init?(coder: NSCoder) {
+    fatalError("init(coder:) has not been implemented")
+  }
+  
+  override func layoutSubviews() {
+    super.layoutSubviews()
+    rootContainer.pin.all()
+    rootContainer.flex.layout()
+  }
+  
+  private func setupConstraints() {
+    addSubview(rootContainer)
+    rootContainer.flex.direction(.row).define { flex in
+      flex.addItem(yearPicker).marginRight(40)
+      flex.addItem(monthPicker)
+    }
+    .alignItems(.center).justifyContent(.center)
+  }
+  
+  func bind(reactor: DatePickerReactor) {    
+    yearPicker.reactor = reactor
+    monthPicker.reactor = reactor
+    reactor.pulse(\.$pickerRow)
+      .compactMap { $0 }
+      .bind(with: self) { owner, value in
+        owner.yearPicker.collectionView.scrollToItem(
+          at: IndexPath(row: value.year, section: 0), at: .top, animated: false
+        )
+        owner.monthPicker.collectionView.scrollToItem(
+          at: IndexPath(row: value.month, section: 0), at: .top, animated: false
+        )
+        owner.yearPicker.collectionView.reloadData()
+        owner.monthPicker.collectionView.reloadData()
+      }
+      .disposed(by: disposeBag)
+  }
+}
+
